+++ Partake/src/main/java/in/partake/model/dao/jpa/JPACalendarLinkageDao.java
@@ -63,7 +63,7 @@ class JPACalendarLinkageDao extends JPADao<CalendarLinkage> implements ICalendar
     }
     
     @Override
-    public long count(PartakeConnection con) throws DAOException {
+    public int count(PartakeConnection con) throws DAOException {
         return countImpl(con, "CalendarLinkages");
     }
